refactor(server): move triggers to nodes/triggers/*.trigger.node.ts; lifecycle-only path (#211)
