+++ apps/server/src/nodes/triggers/base.trigger.node.ts
