+++ apps/server/src/nodes/triggers/debug-tool.trigger.node.ts
