+++ apps/server/src/nodes/triggers/slack.trigger.node.ts
