+++ apps/server/src/tools/call_agent.tool.ts
@@ -3,7 +3,7 @@ import { z } from 'zod';
 import { LangGraphRunnableConfig } from '@langchain/langgraph';
 import { BaseTool } from './base.tool';
 import { LoggerService } from '../services/logger.service';
-import type { Agent } from '../agents/agent';
+import type { Agent } from '../nodes/agent.node';
 import { TriggerMessage } from '../triggers/base.trigger';
 import { BaseMessage } from '@langchain/core/messages';
